+++ Devscord.DiscordFramework/Middlewares/Factories/UserRoleFactory.cs
@@ -2,6 +2,7 @@
 using Devscord.DiscordFramework.Middlewares.Contexts;
 using Discord.Rest;
 using Discord.WebSocket;
+using Watchman.Common.Models;
 
 namespace Devscord.DiscordFramework.Middlewares.Factories
 {
